fix #59

+++ src/cogs/role_distributor.py
@@ -112,22 +112,27 @@ class RoleDistributor(BaseCog):
 
     @commands.Cog.listener()
     async def on_raw_reaction_add(self, payload: discord.RawReactionActionEvent) -> None:
+        member: discord.Member = payload.member
+        if member.bot:
+            return
         mid = payload.message_id
         for rm in self.reaction_messages:
             if rm.mid == mid and rm.ename == payload.emoji.name:
                 guild = self.bot.get_guild(payload.guild_id)
                 role = guild.get_role(rm.rid)
-                await payload.member.add_roles(role)
+                await member.add_roles(role)
                 break
 
     @commands.Cog.listener()
     async def on_raw_reaction_remove(self, payload: discord.RawReactionActionEvent) -> None:
         mid: discord.Message = payload.message_id
+        guild = self.bot.get_guild(payload.guild_id)
+        member = await guild.fetch_member(payload.user_id)
+        if member.bot:
+            return
         for rm in self.reaction_messages:
             if rm.mid == mid and rm.ename == payload.emoji.name:
-                guild = self.bot.get_guild(payload.guild_id)
                 role = guild.get_role(rm.rid)
-                member = await guild.fetch_member(payload.user_id)
                 await member.remove_roles(role)
                 break
